+++ src/main/java/com/daemawiki/daemawiki/domain/mail/event/handler/UserMailSendEventHandler.java
@@ -1,0 +1,7 @@
+package com.daemawiki.daemawiki.domain.mail.event.handler;
+
+import com.daemawiki.daemawiki.domain.mail.event.model.MailSendEvent;
+
+public interface UserMailSendEventHandler {
+    void handleEvent(MailSendEvent event);
+}
