+++ src/main/java/com/daemawiki/daemawiki/domain/mail/service/UserMailSendService.java
@@ -3,8 +3,8 @@ package com.daemawiki.daemawiki.domain.mail.service;
 import com.daemawiki.daemawiki.domain.mail.auth_code.model.AuthCodeModel;
 import com.daemawiki.daemawiki.domain.mail.auth_code.repository.AuthCodeRepository;
 import com.daemawiki.daemawiki.domain.mail.model.MailType;
-import com.daemawiki.daemawiki.domain.mail.model.event.MailSendEvent;
-import com.daemawiki.daemawiki.domain.mail.usecase.UserMailSendEventHandler;
+import com.daemawiki.daemawiki.domain.mail.event.model.MailSendEvent;
+import com.daemawiki.daemawiki.domain.mail.event.handler.UserMailSendEventHandler;
 import com.daemawiki.daemawiki.domain.mail.usecase.UserMailSendUseCase;
 import com.daemawiki.daemawiki.domain.user.repository.UserRepository;
 import com.daemawiki.daemawiki.global.utils.AuthCodeGenerator;
