--- src/main/java/com/daemawiki/daemawiki/domain/mail/usecase/UserMailSendEventHandler.java
@@ -1,7 +1,0 @@
-package com.daemawiki.daemawiki.domain.mail.usecase;
-
-import com.daemawiki.daemawiki.domain.mail.model.event.MailSendEvent;
-
-public interface UserMailSendEventHandler {
-    void handleEvent(MailSendEvent event);
-}
